requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.2.0
django-filter==23.5
channels==4.0.0
channels-redis==4.1.0

